cursor environment deployment page refactor
