+++ apps/webservice/src/app/[workspaceSlug]/(app)/(deploy)/(raw)/systems/[systemSlug]/(raw)/environments/[environmentId]/deployments/_components/SkeletonRow.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { Skeleton } from "@ctrlplane/ui/skeleton";
+import { TableCell, TableRow } from "@ctrlplane/ui/table";
+
+export const SkeletonRow: React.FC = () => (
+  <TableRow className="h-12">
+    {Array.from({ length: 8 }).map((_, index) => (
+      <TableCell key={index}>
+        <Skeleton className="h-4 w-20" />
+      </TableCell>
+    ))}
+  </TableRow>
+);
